Extract Workbench performance API module
